% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_rx}
\alias{scale_color_rx}
\title{Scale color ggplot object}
\usage{
scale_color_rx(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Scale color ggplot object
}
