% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_fill_rx}
\alias{scale_fill_rx}
\title{Scale fill color for ggplot object}
\usage{
scale_fill_rx(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Scale fill color for ggplot object
}
