% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_rx}
\alias{theme_rx}
\title{Function to apply RevelX Theme to a ggplot object}
\usage{
theme_rx(base_size = 12, base_family = "Lato", ...)
}
\arguments{
\item{base_family}{}
}
\value{

}
\description{
Function to apply RevelX Theme to a ggplot object
}
