% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_rx}
\alias{theme_rx}
\title{Function to apply RevelX Theme to a ggplot object}
\usage{
theme_rx(base_size = 12, base_family = "Lato",
  grid_lines = "#eeeeee", angle = 0, background_color = "white",
  legend_position = "bottom", ...)
}
\arguments{
\item{base_size}{Set the base size, default is 12}

\item{base_family}{Set Font familiy , default is Lato}

\item{grid_lines}{Set gridline color, default is NA. Color can be a valid hex value.}

\item{angle}{Set angle of x axis labels}

\item{background_color}{Set background color, default is white. Color can be a valid hex value.}

\item{legend_position}{Set legend position, default is bottom}

\item{...}{}
}
\value{

}
\description{
Function to apply RevelX Theme to a ggplot object
}
